chore: release 0.1.8
升级核心依赖：pyo3 0.25→0.28、numpy 0.25→0.28、polars 0.46→0.53、
strum/strum_macros 0.26→0.28、hashbrown 0.15→0.17、rayon 1.10→1.12，
并适配新版 API（DataFrame::new_infer_height、Column::from、py.detach 等）。

+++ src/core/daily_performance.rs
@@ -504,7 +504,7 @@ mod tests {
         let dp252 = daily_performance(&returns, Some(252)).unwrap();
         let dp365 = daily_performance(&returns, Some(365)).unwrap();
         // annual_returns = mean * yearly_days, so ratio should be 365/252
-        let ratio = dp365.annual_returns as f64 / dp252.annual_returns as f64;
+        let ratio = dp365.annual_returns / dp252.annual_returns;
         assert!((ratio - 365.0 / 252.0).abs() < 0.01);
     }
